requirements.txt
flask
pywebview
playwright
icalendar
requests
python-dotenv
pyyaml
anthropic
openai
google-genai
websockets<14  # google-genai 1.68.x incompatible with websockets 14+ (websockets.client removed)
